chore(architecture): Move rusoto and http utils
I need the `rusoto` module used by AWS sinks (previously in
`sinks::util:rusoto`) in the new `aws_s3` source. This pulls it up into
`src/` alongside other shared utility modules.

As part of moving it up, I also had to move up the `HttpClient` type in
the `sinks::util::http` module, which I pulled into an `http` module.

I followed the existing naming conventions for modules at the top-level
(like `kafka`, `dns`), but I admit that having an `http` module is a bit
confusing in that it conflicts with the `http` crate. I'm open to
suggestions here.
