+++ src/sinks/aws_kinesis_firehose.rs
@@ -3,10 +3,10 @@ use crate::{
     dns::Resolver,
     event::Event,
     region::RegionOrEndpoint,
+    rusoto,
     sinks::util::{
         encoding::{EncodingConfig, EncodingConfiguration},
         retries::RetryLogic,
-        rusoto,
         sink::Response,
         BatchConfig, BatchSettings, Compression, EncodedLength, TowerRequestConfig, VecBuffer,
     },
