+++ src/sinks/elasticsearch.rs
@@ -2,13 +2,15 @@ use crate::{
     config::{DataType, SinkConfig, SinkContext, SinkDescription},
     emit,
     event::Event,
+    http::HttpClient,
     internal_events::{ElasticSearchEventReceived, ElasticSearchMissingKeys},
     region::{region_from_endpoint, RegionOrEndpoint},
+    rusoto,
     sinks::util::{
         encoding::{EncodingConfigWithDefault, EncodingConfiguration},
-        http::{BatchedHttpSink, HttpClient, HttpSink},
+        http::{BatchedHttpSink, HttpSink},
         retries::{RetryAction, RetryLogic},
-        rusoto, BatchConfig, BatchSettings, Buffer, Compression, TowerRequestConfig,
+        BatchConfig, BatchSettings, Buffer, Compression, TowerRequestConfig,
     },
     template::{Template, TemplateError},
     tls::{TlsOptions, TlsSettings},
