+++ src/sinks/loki.rs
@@ -15,10 +15,11 @@
 use crate::{
     config::{log_schema, DataType, GenerateConfig, SinkConfig, SinkContext, SinkDescription},
     event::{self, Event, Value},
+    http::HttpClient,
     sinks::util::{
         buffer::loki::{LokiBuffer, LokiEvent, LokiRecord},
         encoding::{EncodingConfigWithDefault, EncodingConfiguration},
-        http::{Auth, BatchedHttpSink, HttpClient, HttpSink},
+        http::{Auth, BatchedHttpSink, HttpSink},
         BatchConfig, BatchSettings, TowerRequestConfig, UriSerde,
     },
     template::Template,
